requirements.txt
# DevMentor AI - Python Dependencies

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0

# Distributed Training
deepspeed>=0.9.0
bitsandbytes>=0.40.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
datasets>=2.12.0

# Web Framework (API Server)
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Async/HTTP
aiohttp>=3.8.0
httpx>=0.24.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and Logging
wandb>=0.15.0
tensorboard>=2.13.0
mlflow>=2.4.0
prometheus-client>=0.17.0

# Code Analysis
ast-comments>=1.1.0
libcst>=1.0.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.80.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0
pylint>=2.17.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.0
mkdocs-material>=9.1.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
notebook>=6.5.0

# Optional: GPU acceleration
# nvidia-ml-py3>=7.352.0  # For GPU monitoring
# flash-attn>=2.0.0  # Requires compilation, optional

# Optional: Advanced optimizations
# apex  # NVIDIA Apex (requires compilation)
# triton>=2.0.0  # For custom kernels
